+++ pyobjc-framework-Metal/PyObjCTest/test_mtlargumentencoder.py
@@ -67,6 +67,12 @@ class TestMTLArgumentEncoderHelper(Metal.NSObject):
     def setIntersectionFunctionTables_withRange_(self, a, b):
         pass
 
+    def setDepthStencilState_atIndex_(self, a, b):
+        pass
+
+    def setDepthStencilStates_withRange_(self, a, b):
+        pass
+
     def setComputePipelineState_atIndex_(self, a, b):
         pass
 
@@ -253,6 +259,26 @@ class TestMTLArgumentEncoder(TestCase):
         )
         self.assertArgHasType(
             TestMTLArgumentEncoderHelper.setIntersectionFunctionTables_withRange_,
+            1,
+            Metal.NSRange.__typestr__,
+        )
+
+        self.assertArgHasType(
+            TestMTLArgumentEncoderHelper.setDepthStencilState_atIndex_,
+            1,
+            objc._C_NSUInteger,
+        )
+
+        self.assertArgHasType(
+            TestMTLArgumentEncoderHelper.setDepthStencilStates_withRange_,
+            0,
+            b"n^@",
+        )
+        self.assertArgSizeInArg(
+            TestMTLArgumentEncoderHelper.setDepthStencilStates_withRange_, 0, 1
+        )
+        self.assertArgHasType(
+            TestMTLArgumentEncoderHelper.setDepthStencilStates_withRange_,
             1,
             Metal.NSRange.__typestr__,
         )