+++ pyobjc-framework-Metal/PyObjCTest/test_mtldevice.py
@@ -432,6 +432,7 @@ class TestMTLDevice(TestCase):
         self.assertEqual(Metal.MTLGPUFamilyApple7, 1007)
         self.assertEqual(Metal.MTLGPUFamilyApple8, 1008)
         self.assertEqual(Metal.MTLGPUFamilyApple9, 1009)
+        self.assertEqual(Metal.MTLGPUFamilyApple10, 1010)
 
         self.assertEqual(Metal.MTLGPUFamilyMac1, 2001)
         self.assertEqual(Metal.MTLGPUFamilyMac2, 2002)